refactor: 함수명/파일명 수정 - 함수명과 파일명을 일관성 있도록 수정한다.
