+++ src/App.js
@@ -1,17 +1,14 @@
-import prepareCarNames from './controllers/carContoller.js';
-import startGame from './controllers/gameController.js';
-import {
-  getCarNamesInput,
-  getRoundCount,
-} from './controllers/inputController.js';
+import parseCarNames from './controllers/car.js';
+import startGame from './controllers/game.js';
+import { getCarNamesInput, getRoundCountInput } from './controllers/input.js';
 import { validateRoundCount } from './validators/validator.js';
 
 class App {
   async run() {
     const carNamesInput = await getCarNamesInput();
-    const roundCountInput = await getRoundCount();
+    const roundCountInput = await getRoundCountInput();
 
-    const carNames = prepareCarNames(carNamesInput);
+    const carNames = parseCarNames(carNamesInput);
     const roundCount = validateRoundCount(roundCountInput);
 
     await startGame(carNames, roundCount);
